refactor: Isolate vitest helpers from other imports to avoid cjs problems.
