+++ tests/handlers/InputAdded.test.ts
@@ -16,6 +16,7 @@ import {
     NFT,
     Token,
 } from '../../src/model';
+import { mockModelImplementation } from '../stubs/mocks';
 import {
     block,
     ctx,
@@ -25,7 +26,6 @@ import {
     logErc721Transfer,
     logs,
 } from '../stubs/params';
-import { mockModelImplementation } from '../stubs/utils';
 
 vi.mock('../../src/abi/ERC20');
